requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==41.0.7
defusedxml==0.7.1
Django==5.2.13
djangosaml2==1.12.0
elementpath==4.8.0
idna==3.13
lxml==6.1.0
mozilla-django-oidc==5.0.2
ngrok==1.4.0
oauthlib==3.3.1
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==23.2.0
pysaml2==7.5.4
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.33.1
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.9.0
social-auth-core==4.9.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
xmlschema==2.5.1
xmlsec==1.3.17
psycopg2-binary==2.9.12
python-dotenv==1.2.2
